feat: add web login features and move non rest controller to FrontController.
