+++ src/main/resources/templates/login.html
@@ -1,0 +1,69 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <title>회원 로그인</title>
+    <!-- 부트스트랩 CSS -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
+    <!-- Pretendard 한글 폰트 -->
+    <link rel="stylesheet" as="style" crossorigin href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/static/pretendard-dynamic-subset.min.css" />
+    <link rel="stylesheet" href="/style/style.css" />
+</head>
+<body class="bg-body-tertiary">
+    <div class="container text-center position-absolute top-50 start-50 translate-middle" style="width: 400px">
+        <h1>회원 로그인</h1>
+        <form id="loginForm">
+            <div>
+            <label for="username" class="form-label">사용자 이름:</label>
+            <input type="text" id="username" class="form-control" />
+            </div>
+            <div>
+            <label for="password" class="form-label">암호:</label>
+            <input type="password" id="password" class="form-control" />
+            </div>
+            <hr class="my-4" />
+            <div>
+            <button type="submit" class="btn btn-primary w-100">로그인</button>
+            </div>
+        </form>
+
+        <script>
+            document.getElementById('loginForm')
+                .addEventListener('submit', function(event) {
+                    event.preventDefault();
+                    const username = document.getElementById('username').value;
+                    const password = document.getElementById('password').value;
+                    const data = { username, password };
+                    postJSON(data);
+                });
+        </script>
+    </div>
+
+
+    <script>
+    async function postJSON(data) {
+        try {
+            const response = await fetch("/api/login", {
+                    method: "POST",
+                    body: JSON.stringify(data)
+                }
+            );
+            const result = await response.json();
+            if (response.status === 500) {
+                alert("로그인을 실패하였습니다. ID와 비밀번호를 확인하십시오.");
+                return;
+            }
+            console.log("성공:", result);
+            localStorage.setItem("token", result.token);
+            location.href = "/users/"+result.uid+"/profile";   
+        } catch (error) {
+            alert("오류 발생:", error);
+        }
+      }
+    </script>      
+
+    <!-- 부트스트랩용 js -->
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
+    integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
+    crossorigin="anonymous"></script>
+</body>
+</html>
