+++ src/main/resources/templates/register.html
@@ -41,8 +41,29 @@
             <p>가입을 완료하였습니다.</p>
         </div>
     </div>
-
-
+    <script>
+        async function postJSON(data) {
+            try {
+                const formData = new FormData();
+                const blob = new Blob([JSON.stringify(data)], { type: "application/json" });
+                formData.append("user", blob);
+                
+                const response = await fetch("/api/register", {
+                    method: "POST",
+                    body: formData
+                }
+                );
+                
+                console.log( formData );
+                const result = await response.json();
+                console.log("성공:", result);
+            } catch (error) {
+                console.error("실패:", error);
+            }
+        }
+    </script>      
+    
+    
     <!-- 부트스트랩용 js -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
     integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
